Fixed bugs on the rollback migrations and seeders;

+++ database/migrations/2021_06_09_110437_update2_disbursement_vouchers_table.php
@@ -30,7 +30,12 @@ class Update2DisbursementVouchersTable extends Migration
         Schema::table('disbursement_vouchers', function (Blueprint $table) {
             $table->dropForeign('disbursement_vouchers_funding_source_foreign');
             $table->dropColumn('funding_source');
-            $table->dropColumn('uacs_object_code');
         });
+
+        if (Schema::hasColumn('disbursement_vouchers', 'uacs_object_code')) {
+            Schema::table('disbursement_vouchers', function (Blueprint $table) {
+                $table->dropColumn('uacs_object_code');
+            });
+        }
     }
 }
